+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  * Fix django.conf.urls deprecation warnings
  * Added user_handle field to Authenticator model
  * Added DJANGO_FIDO_RESIDENT_KEY (default False) app setting
+ * Added passwordless auth backend
 
 ## 0.36
  * Sort users in admin Authenticator add form by username